+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Updated
 
 - Introduced `AWSExecutor` as a second parent to the `EC2Executor`
+- Updated requirements.txt to pin aws executor plugins to pre-release version 0.1.0rc0
 
 ### Tests