+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [UNRELEASED]
 
+### Added
+
+-  Logic to specify that only the base covalent-aws-plugins package is to be installed.
+
 ### Updated
 
 - Made EC2 executor async aware