+++ CHANGELOG.md
@@ -13,6 +13,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Updated
 
+## [0.5.1] - 2022-09-29
+
+### Fixed
+
+- Typo in `release.yml`
+
 ## [0.5.0] - 2022-09-29
 
 ### Added